January/28th-Jan-Wed/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pymysql>=1.1.2",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.uv.workspace]
members = [
    "fastapi_relationship",
]
